refactor: View 컴포넌트를 `components/view/View` 에서 `components/View`로 이동
