+++ app/(tabs)/(onboard)/agreement.tsx
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
 import Text from '@/components/Text';
 import { TextSize } from '@/enums/TextSize';
-import View from '@/components/views/View';
+import View from '@/components/View';
 import Button from '@/components/Button';
 import Checkbox from '@/components/Checkbox';
 import useAgreementStore from '@/stores/agreementStore';
