+++ app/(tabs)/(onboard)/signup.tsx
@@ -1,7 +1,7 @@
 import {Keyboard, Pressable, StyleSheet, useColorScheme} from 'react-native';
 import Text from '@/components/Text';
 import { TextSize } from '@/enums/TextSize';
-import View from '@/components/views/View';
+import View from '@/components/View';
 import Button from '@/components/Button';
 import Textfield from '@/components/Textfield';
 import { useEffect, useState } from 'react';
