+++ app/(tabs)/index.tsx
@@ -1,5 +1,5 @@
 import { Image, SafeAreaView, StyleSheet } from 'react-native';
-import View from '@/components/views/View';
+import View from '@/components/View';
 import Text from '@/components/Text';
 import { TextSize } from '@/enums/TextSize';
 import Button from '@/components/Button';
